requirements.txt
python-telegram-bot
redis
python-dotenv
PyYAML
pandas
